FlaskServer/requirements.txt
Flask
ultralytics
numpy
opencv-python
Pillow